requirements.txt
# Usage: pip install -r requirements.txt

numpy>=1.24.4
opencv-python>=4.8.0
Pillow>=9.4.0
PyYAML>=6.0.1
torch>=2.0.1
torchvision>=0.15.2